+++ io.sarl.jaak.kernel/src/main/java/io/sarl/jaak/environment/external/frustum/TurtleFrustum.java
@@ -23,6 +23,7 @@ import io.sarl.jaak.environment.external.EnvironmentArea;
 
 import java.util.Iterator;
 
+import org.arakhne.afc.math.continous.object2d.Point2f;
 import org.arakhne.afc.math.discrete.object2d.Point2i;
 
 /** This interface defines a frustum for for a turtle.
@@ -42,6 +43,6 @@ public interface TurtleFrustum {
 	 * @param environment is the environment in which the frustum should perceive.
 	 * @return the iterator on perceived cells' positions.
 	 */
-	Iterator<Point2i> getPerceivedCells(Point2i origin, float direction, EnvironmentArea environment);
+	Iterator<Point2f> getPerceivedCells(Point2f origin, float direction, EnvironmentArea environment);
 
 }
